requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
fastapi
pandas
mysql-connector-python
python-dotenv
pypdf